+++ packages/webpack/config/react-js.js
@@ -1,0 +1,14 @@
+// Local import
+const { srcDir } = require('./path');
+
+module.exports = {
+  module: {
+    rules: [
+      {
+        test: /\.jsx?$/,
+        include: srcDir,
+        use: 'babel-loader'
+      }
+    ]
+  }
+};