+++ packages/webpack/scripts/watch.js
@@ -9,18 +9,18 @@ const webpack = require('webpack');
 const WebpackDevServer = require('webpack-dev-server');
 
 // Local import
-const logger = require('./logger');
 const { env, host, port } = require('../config/env');
 const { staticDir } = require('../config/path');
 const webpackConfig = require('../webpack.config');
+const logger = require('./logger');
 
-module.exports = (mode, options) => {
+module.exports = options => {
   // Initialize console
   clearConsole();
   logger.start(`Starting build in ${env} mode`);
 
   // Set DevServer
-  const devConfig = webpackConfig(env, { mode, ...options });
+  const devConfig = webpackConfig(env, options);
   const compiler = webpack(devConfig);
   const devOptions = {
     contentBase: staticDir,