+++ packages/webpack/webpack.config.js
@@ -6,20 +6,20 @@ const { smart } = require('webpack-merge');
 const commonConfig = require('./webpack.config.common');
 
 module.exports = (env, options) => {
-  const { mode, add } = options;
-  const modeConfig = require(`./config/${mode}`);
-
+  const { add, lang, mode } = options;
+  const targetConfig = require(`./config/${mode}-${lang}`);
   let additionalConfig = {};
+
   if (add) additionalConfig = require(resolve(process.cwd(), add));
 
   switch (env) {
     case 'development': {
       const devConfig = require('./webpack.config.dev');
-      return smart(commonConfig, modeConfig, devConfig, additionalConfig);
+      return smart(commonConfig, targetConfig, devConfig, additionalConfig);
     }
     case 'production': {
       const prodConfig = require('./webpack.config.prod');
-      return smart(commonConfig, modeConfig, prodConfig, additionalConfig);
+      return smart(commonConfig, targetConfig, prodConfig, additionalConfig);
     }
     case 'test':
     default: