+++ packages/webpack/CHANGELOG.md
@@ -2,5 +2,6 @@
 
 ## 0.5.4 (Apr 23, 2018)
 
+* Mode `mode` and `lang` params into option
 * Add `chunksSortMode` option into `html-webpack-plugin`
 * Replace `DefinePlugin` into `DotenvPlugin` at common config